Add virtual device type for all devices

+++ test/lua_manager_dependencies.cpp
@@ -11,28 +11,28 @@ void LuaManagerTest::SetUp()
 	lua_hooks.push_back(subhook_new((void *) tolua_tostring,        (void *) mock_tolua_tostring,       SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) tolua_tonumber,        (void *) mock_tolua_tonumber,       SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) luaL_loadstring,       (void *) mock_luaL_loadstring,      SUBHOOK_64BIT_OFFSET));
-	lua_hooks.push_back(subhook_new((void *) tolua_tousertype,      (void *) mock_tolua_tousertype,     SUBHOOK_64BIT_OFFSET));	
+	lua_hooks.push_back(subhook_new((void *) tolua_tousertype,      (void *) mock_tolua_tousertype,     SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) luaL_loadfile,         (void *) mock_luaL_loadfile,        SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) luaL_newstate,         (void *) mock_luaL_newstate,        SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) lua_gc,                (void *) mock_lua_gc,               SUBHOOK_64BIT_OFFSET));
-	lua_hooks.push_back(subhook_new((void *) luaL_openlibs,         (void *) mock_luaL_openlibs,        SUBHOOK_64BIT_OFFSET));	
+	lua_hooks.push_back(subhook_new((void *) luaL_openlibs,         (void *) mock_luaL_openlibs,        SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) tolua_PAC_dev_open,    (void *) mock_tolua_PAC_dev_open,   SUBHOOK_64BIT_OFFSET));
 	lua_hooks.push_back(subhook_new((void *) tolua_IOT_dev_open,    (void *) mock_tolua_IOT_dev_open,   SUBHOOK_64BIT_OFFSET));
-	lua_hooks.push_back(subhook_new((void *) check_file,            (void *) mock_check_file,           SUBHOOK_64BIT_OFFSET));    
+	lua_hooks.push_back(subhook_new((void *) check_file,            (void *) mock_check_file,           SUBHOOK_64BIT_OFFSET));
     lua_hooks.push_back(subhook_new((void *) G_TECH_OBJECT_MNGR,    (void *) mock_G_TECH_OBJECT_MNGR,   SUBHOOK_64BIT_OFFSET));
     lua_hooks.push_back(subhook_new((void *) lua_tolstring,         (void *) mock_lua_tolstring,        SUBHOOK_64BIT_OFFSET));
     lua_hooks.push_back(subhook_new((void *) lua_settop,            (void *) mock_lua_settop,           SUBHOOK_64BIT_OFFSET));
-	    
-	// Install hooks 
-	for (int i = 0; i < lua_hooks.size(); i++) {
+
+	// Install hooks
+	for (size_t i = 0; i < lua_hooks.size(); i++) {
 		subhook_install(lua_hooks[i]);
 	}
 }
 
 void LuaManagerTest::TearDown()
 {
-	// Remove the hooks and free memory 
-	for (int i = 0; i < lua_hooks.size(); i++) {
+	// Remove the hooks and free memory
+	for (size_t i = 0; i < lua_hooks.size(); i++) {
 		subhook_remove(lua_hooks[i]);
 		subhook_free(lua_hooks[i]);
 	}
